Enum oplog::Operation

pub enum Operation {
    Noop {
        id: i64,
        timestamp: DateTime<UTC>,
        message: String,
    },
    Insert {
        id: i64,
        timestamp: DateTime<UTC>,
        namespace: String,
        document: Document,
    },
    Update {
        id: i64,
        timestamp: DateTime<UTC>,
        namespace: String,
        query: Document,
        update: Document,
    },
    Delete {
        id: i64,
        timestamp: DateTime<UTC>,
        namespace: String,
        query: Document,
    },
    Command {
        id: i64,
        timestamp: DateTime<UTC>,
        namespace: String,
        command: Document,
    },
}

A MongoDB oplog operation.

Variants
Noop

A no-op as inserted periodically by MongoDB or used to initiate new replica sets.

Fields

id: i64

A unique identifier for this operation.

timestamp: DateTime<UTC>

The time of the operation.

message: String

The message associated with this operation.

Insert

An insert of a document into a specific database and collection.

Fields

id: i64

A unique identifier for this operation.

timestamp: DateTime<UTC>

The time of the operation.

namespace: String

The full namespace of the operation including its database and collection.

document: Document

The BSON document inserted into the namespace.

Update

An update of a document in a specific database and collection matching a given query.

Fields

id: i64

A unique identifier for this operation.

timestamp: DateTime<UTC>

The time of the operation.

namespace: String

The full namespace of the operation including its database and collection.

query: Document

The BSON selection criteria for the update.

update: Document

The BSON update applied in this operation.

Delete

The deletion of a document in a specific database and collection matching a given query.

Fields

id: i64

A unique identifier for this operation.

timestamp: DateTime<UTC>

The time of the operation.

namespace: String

The full namespace of the operation including its database and collection.

query: Document

The BSON selection criteria for the delete.

Command

A command such as the creation or deletion of a collection.

Fields

id: i64

A unique identifier for this operation.

timestamp: DateTime<UTC>

The time of the operation.

namespace: String

The full namespace of the operation including its database and collection.

command: Document

The BSON command.

Methods

impl Operation

fn new(document: &Document) -> Result<Operation>

Try to create a new Operation from a BSON document.